move features components from compontents dir to features dir
