+++ src/pages/main-page/main-page.tsx
@@ -6,7 +6,7 @@ import { useAddLocalFilesOrFolders } from '@hooks/use-add-local-files-folders';
 import { useAppStore } from '@store/app-store';
 import { TabsPane } from '@features/tabs-pane';
 import { TabView } from '@features/tab-view/tab-view';
-import { StartGuide } from '@components/start-guide';
+import { StartGuide } from '@features/start-guide';
 import { createSQLScript } from '@controllers/sql-script';
 import { getOrCreateTabFromScript } from '@controllers/tab';
 import { importSQLFiles } from '@utils/import-script-file';
@@ -17,7 +17,6 @@ export const MainPage = () => {
   /**
    * Common hooks
    */
-
   const { handleAddFile, handleAddFolder } = useAddLocalFilesOrFolders();
   const { colorScheme } = useMantineColorScheme();
   const [layoutSizes, setOuterLayoutSizes] = useLocalStorage<number[]>({
